ai agent test/velocity-logic-agent/requirements.txt
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pandas>=2.0.0
fpdf2>=2.7.0
python-dotenv>=1.0.0
thefuzz>=0.19.0
python-Levenshtein>=0.21.0
flask>=3.0.0

